requirements.txt
Flask==3.0.2
requests==2.31.0
google-generativeai==0.3.2
Flask-Session==0.6.0
Pillow==10.2.0
wikipedia-api==0.6.0
python-dateutil==2.8.2 